Изменения визуала + микрофиксы

+++ src/adminBot/scenes/controllers/procedureMethods.js
@@ -1,4 +1,4 @@
-const { Procedure } = require('../../../database/models/index');
+const {Procedure} = require('../../../database/models/index');
 
 class ProcedureMethods {
     /**
@@ -9,7 +9,7 @@ class ProcedureMethods {
      */
 
     static async enterRussianName(ctx) {
-        const { session, callbackQuery } = ctx;
+        const {session, callbackQuery} = ctx;
         session.editingProcedure = {};
         const selectGroup = callbackQuery.data.split('_')[3];
 
@@ -30,8 +30,6 @@ class ProcedureMethods {
         const newProcedureNumber = lastNumber + 1;
 
         session.editingProcedure.englishName = `${selectGroup}${newProcedureNumber}`;
-        console.log(session.editingProcedure.englishName);
-
         session.tempMessage = await ctx.reply('Введите название процедуры на русском:');
         return ctx.wizard.next();
     }
@@ -44,7 +42,7 @@ class ProcedureMethods {
      */
 
     static async enterDuration(ctx) {
-        const { session, message } = ctx;
+        const {session, message} = ctx;
 
         if (!message.text || message.text.trim().length === 0) {
             await ctx.reply(
@@ -63,12 +61,11 @@ class ProcedureMethods {
 
     /**
      * Создание новой процедуры - шаг 3.
-     * Сохранение новой процедуры в базе данных.
+     * Ввод цены процедуры
      * @param {object} ctx - Контекст Telegram.
      */
-
-    static async saveProcedure(ctx) {
-        const { session, message } = ctx;
+    static async enterPrice(ctx) {
+        const {session, message} = ctx;
         const procedureTime = parseFloat(message.text);
 
         if (isNaN(procedureTime) || procedureTime < 0.5 || procedureTime > 12) {
@@ -82,8 +79,32 @@ class ProcedureMethods {
         await ctx.deleteMessage(message.message_id);
         await ctx.deleteMessage(session.tempMessage.message_id);
 
-        const { englishName, russianName, duration } = session.editingProcedure;
-        const newProcedure = new Procedure({ englishName, russianName, duration });
+        session.tempMessage = await ctx.reply('Введите цену процедуры в рублях:');
+        return ctx.wizard.next();
+    }
+
+    /**
+     * Создание новой процедуры - шаг 4.
+     * Сохранение процедуры в базе данных.
+     * @param {object} ctx - Контекст Telegram.
+     */
+    static async saveProcedure(ctx) {
+        const {session, message} = ctx;
+        const procedurePrice = parseFloat(message.text);
+
+        if (isNaN(procedurePrice) || procedurePrice <= 0) {
+            await ctx.reply(
+                'Ошибка: цена должна быть положительным числом. Пожалуйста, введите цену снова.'
+            );
+            return ctx.wizard.selectStep(ctx.wizard.cursor - 1);
+        }
+
+        session.editingProcedure.price = procedurePrice;
+        await ctx.deleteMessage(message.message_id);
+        await ctx.deleteMessage(session.tempMessage.message_id);
+
+        const {englishName, russianName, duration, price} = session.editingProcedure;
+        const newProcedure = new Procedure({englishName, russianName, duration, price});
         await newProcedure.save();
 
         session.tempMessage = await ctx.reply(`Новая процедура "${russianName}" успешно создана.`);
@@ -93,29 +114,33 @@ class ProcedureMethods {
 
     /**
      * Редактирование процедуры - шаг 1.
-     * Получение текущей длительности процедуры.
+     * Получение текущей длительности и цены процедуры.
      * @param {object} ctx - Контекст Telegram.
      */
 
-    static async enterNewProcedureDuration(ctx) {
-        const { session, callbackQuery } = ctx;
+    static async enterNewProcedureDetails(ctx) {
+        const {session, callbackQuery} = ctx;
         session.editingProcedure = {};
         session.editingProcedure.englishName = callbackQuery.data.split('_')[3];
-        const procedure = await Procedure.findOne({ englishName: session.editingProcedure.englishName });
+
+        const procedure = await Procedure.findOne({englishName: session.editingProcedure.englishName});
+
         session.tempMessage = await ctx.reply(
-            `Текущая длительность процедуры "${procedure.russianName}": ${procedure.duration} часа.\nВведите новую длительность процедуры "${procedure.russianName}" в часах:`
+            `Текущая длительность процедуры "${procedure.russianName}": ${procedure.duration} часа.\n` +
+            `Текущая цена: ${procedure.price} руб.\nВведите новую длительность процедуры "${procedure.russianName}" в часах:`
         );
+
         return ctx.wizard.next();
     }
 
     /**
      * Редактирование процедуры - шаг 2.
-     * Обновление длительности процедуры.
+     * Обновление длительности процедуры и запрос новой цены.
      * @param {object} ctx - Контекст Telegram.
      */
 
-    static async saveUpdatedProcedure(ctx) {
-        const { session, message } = ctx;
+    static async enterNewPrice(ctx) {
+        const {session, message} = ctx;
         const procedureTime = parseFloat(message.text);
 
         if (isNaN(procedureTime) || procedureTime < 0.5 || procedureTime > 12) {
@@ -125,22 +150,49 @@ class ProcedureMethods {
             return ctx.wizard.selectStep(ctx.wizard.cursor - 1);
         }
 
-        const procedure = await Procedure.findOne({ englishName: session.editingProcedure.englishName });
-        const newDuration = procedureTime;
+        session.editingProcedure.newDuration = procedureTime;
+        await ctx.deleteMessage(message.message_id);
+        await ctx.deleteMessage(session.tempMessage.message_id);
 
-        procedure.duration = newDuration;
+        session.tempMessage = await ctx.reply('Введите новую цену процедуры в рублях:');
+        return ctx.wizard.next();
+    }
+
+    /**
+     * Редактирование процедуры - шаг 3.
+     * Сохранение новой длительности и цены процедуры.
+     * @param {object} ctx - Контекст Telegram.
+     */
+
+    static async saveUpdatedProcedure(ctx) {
+        const {session, message} = ctx;
+        const newPrice = parseFloat(message.text);
+
+        if (isNaN(newPrice) || newPrice <= 0) {
+            await ctx.reply(
+                'Ошибка: цена должна быть положительным числом. Пожалуйста, введите цену снова.'
+            );
+            return ctx.wizard.selectStep(ctx.wizard.cursor - 1);
+        }
+
+        const procedure = await Procedure.findOne({englishName: session.editingProcedure.englishName});
+
+        procedure.duration = session.editingProcedure.newDuration;
+        procedure.price = newPrice;
         await procedure.save();
 
         await ctx.deleteMessage(message.message_id);
         await ctx.deleteMessage(session.tempMessage.message_id);
 
         session.tempMessage = await ctx.reply(
-            `Длительность процедуры "${procedure.russianName}" успешно изменена на ${procedure.duration} часа.`
+            `Длительность процедуры "${procedure.russianName}" успешно изменена на ${procedure.duration} часа, ` +
+            `а цена на ${procedure.price} руб.`
         );
         setTimeout(() => ctx.deleteMessage(session.tempMessage.message_id), 5000);
 
         return ctx.scene.leave();
     }
+
 }
 
 module.exports = ProcedureMethods;
